Fix a couple of potential error conditions.
Added some additional checks on the YouTube API JSON feed data for cases where the YouTube API omits certain tags when the tag content would be zero or blank. For example, a newly uploaded video with no views would cause a Bail Out error...instead of including the view count tag with 0, the YouTube API doesn't include that tag in the feed.